+++ hel/tests/test_views.py
@@ -174,6 +174,15 @@ class PkgSearchTests(unittest.TestCase):
             ('Kjers', [self.pkg2],)
         ]
 
+    @param('owners')
+    def test_pkg_search_owners(self):
+        return [
+            (['root'], [self.pkg1, self.pkg2, self.pkg3],),
+            (['test'], [],),
+            (['root', 'test'], [],),
+            (['root', 'crackes'], [self.pkg1],)
+        ]
+
     @one_value_param('screen_desc')
     def test_pkg_search_screen_desc(self):
         return [
@@ -249,6 +258,15 @@ class PkgSearchTests(unittest.TestCase):
               'http://img.example.com/img23'], [self.pkg2],),
             (['http://img.example.com/img42'], [],),
             (['http://img.example.com/img32'], [self.pkg3],)
+        ]
+
+    @one_value_param('q')
+    def test_pkg_search_q(self):
+        return [
+            ('root', [self.pkg1, self.pkg2, self.pkg3],),
+            ('Change package "My first"', [self.pkg1],),
+            ('hello people xD', [],),
+            ('test img-2 2 21.', [self.pkg2],)
         ]
 
     def test_bad_search_param(self):