+++ CHANGES.rst
@@ -1,3 +1,10 @@
+3.1.0
+-----
+- Added ``owners`` search param.
+- Added ``q`` search param.
+- Added group ``~banned``.
+- Wrapped ``User:PATCH``.
+
 3.0.3
 -----
 - Added dependency remove code.